pyproject.toml
[project]
name = "crypto-decision-support"
version = "0.1.0"
description = "System wspomagania decyzji inwestycyjnych na rynku kryptowalut z wykorzystaniem ML"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name="Your Name"}]

dependencies = [
  "numpy>=1.26",
  "pandas>=2.2.0",
  "polars>=1.5.0",
  "scikit-learn>=1.4.0",
  "xgboost>=2.0.0",
  "lightgbm>=4.0.0",
  "optuna>=3.6.0",
  "torch>=2.2.0",
  "pytorch-lightning>=2.3.0",
  "torchmetrics>=1.3.0",
  "shap>=0.44.0",
  "ccxt>=4.2.0",
  "pyyaml>=6.0.1",
  "plotly>=5.20.0",
  "streamlit>=1.37.0",
  "mlflow>=2.14.0",
  "loguru>=0.7.0",
  "rich>=13.7.0",
  "joblib>=1.3.0",
  "matplotlib>=3.8.0"
]

[project.optional-dependencies]
backtest = ["vectorbt>=0.25.0","numba>=0.59.0"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E","F","I","UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
strict = false
